Split PlotView into PlotCanvas and PlotView
