+++ bokehjs/src/lib/core/ui_events.ts
@@ -9,7 +9,7 @@ import {getDeltaY} from "./util/wheel"
 import {reversed, is_empty} from "./util/array"
 import {isString} from "./util/types"
 import {is_mobile} from "./util/compat"
-import {PlotView} from "../models/plots/plot"
+import {PlotCanvasView} from "../models/plots/plot_canvas"
 import {ToolView} from "../models/tools/tool"
 import {ContextMenu} from "./util/menus"
 import {RendererView} from "../models/renderers/renderer"
@@ -282,7 +282,7 @@ export class UIEventBus implements EventListenerObject {
     }
   }
 
-  protected _hit_test_renderers(plot_view: PlotView, sx: number, sy: number): RendererView | null {
+  protected _hit_test_renderers(plot_view: PlotCanvasView, sx: number, sy: number): RendererView | null {
     const views = plot_view.get_renderer_views()
 
     for (const view of reversed(views)) {
@@ -300,15 +300,15 @@ export class UIEventBus implements EventListenerObject {
     this.hit_area.style.cursor = cursor
   }
 
-  protected _hit_test_frame(plot_view: PlotView, sx: number, sy: number): boolean {
+  protected _hit_test_frame(plot_view: PlotCanvasView, sx: number, sy: number): boolean {
     return plot_view.frame.bbox.contains(sx, sy)
   }
 
-  protected _hit_test_canvas(plot_view: PlotView, sx: number, sy: number): boolean {
+  protected _hit_test_canvas(plot_view: PlotCanvasView, sx: number, sy: number): boolean {
     return plot_view.layout.bbox.contains(sx, sy)
   }
 
-  protected _hit_test_plot(sx: number, sy: number): PlotView | null {
+  protected _hit_test_plot(sx: number, sy: number): PlotCanvasView | null {
     // TODO: z-index
     for (const plot_view of this.canvas_view.plot_views) {
       if (plot_view.layout.bbox.contains(sx, sy))
@@ -318,11 +318,11 @@ export class UIEventBus implements EventListenerObject {
     return null
   }
 
-  protected _prev_move: {sx: number, sy: number, plot_view: PlotView | null} | null = null
+  protected _prev_move: {sx: number, sy: number, plot_view: PlotCanvasView | null} | null = null
 
-  protected _curr_pan: {plot_view: PlotView} | null = null
-  protected _curr_pinch: {plot_view: PlotView} | null = null
-  protected _curr_rotate: {plot_view: PlotView} | null = null
+  protected _curr_pan: {plot_view: PlotCanvasView} | null = null
+  protected _curr_pinch: {plot_view: PlotCanvasView} | null = null
+  protected _curr_rotate: {plot_view: PlotCanvasView} | null = null
 
   _trigger<E extends UIEvent>(signal: UISignal<E>, e: E, srcEvent: Event): void {
     const {sx, sy} = e
@@ -330,7 +330,7 @@ export class UIEventBus implements EventListenerObject {
     const curr_view = plot_view
 
     if (e.type == "panstart" || e.type == "pan" || e.type == "panend") {
-      let pan_view: PlotView | null
+      let pan_view: PlotCanvasView | null
       if (e.type == "panstart" && curr_view != null) {
         this._curr_pan = {plot_view: curr_view}
         pan_view = curr_view
@@ -347,7 +347,7 @@ export class UIEventBus implements EventListenerObject {
         this.__trigger(pan_view, signal, e, srcEvent)
       }
     } else if (e.type == "pinchstart" || e.type == "pinch" || e.type == "pinchend") {
-      let pinch_view: PlotView | null
+      let pinch_view: PlotCanvasView | null
       if (e.type == "pinchstart" && curr_view != null) {
         this._curr_pinch = {plot_view: curr_view}
         pinch_view = curr_view
@@ -364,7 +364,7 @@ export class UIEventBus implements EventListenerObject {
         this.__trigger(pinch_view, signal, e, srcEvent)
       }
     } else if (e.type == "rotatestart" || e.type == "rotate" || e.type == "rotateend") {
-      let rotate_view: PlotView | null
+      let rotate_view: PlotCanvasView | null
       if (e.type == "rotatestart" && curr_view != null) {
         this._curr_rotate = {plot_view: curr_view}
         rotate_view = curr_view
@@ -403,7 +403,7 @@ export class UIEventBus implements EventListenerObject {
     }
   }
 
-  __trigger<E extends UIEvent>(plot_view: PlotView, signal: UISignal<E>, e: E, srcEvent: Event): void {
+  __trigger<E extends UIEvent>(plot_view: PlotCanvasView, signal: UISignal<E>, e: E, srcEvent: Event): void {
     const gestures = plot_view.model.toolbar.gestures
     type BaseType = keyof typeof gestures
 
@@ -492,7 +492,7 @@ export class UIEventBus implements EventListenerObject {
     signal.emit({id, e})
   }
 
-  protected _trigger_bokeh_event(plot_view: PlotView, e: UIEvent): void {
+  protected _trigger_bokeh_event(plot_view: PlotCanvasView, e: UIEvent): void {
     const ev = (() => {
       const {sx, sy} = e
       const x = plot_view.frame.x_scale.invert(sx)
